squash! Update install-utils.js
Revert the erroneous change to the `requireChmod` function, and remove it entirely.

It seems that the logic of the method was originally correct, and the actual root cause is `fs.access` failing to report missing access permissions correctly when run inside a container; `err` is always `null`, even when the file clearly does not grant permission to the user. This seems to happen with read, write and execute, for any file or directory.

The `chmod` function seems to works regardless of whether the file exists or has the permissions already set. So the `fs.access` check is redundant and can be removed.